feat: move trace log component code to its own package
